requirements.txt
flask==2.3.3
crewai==0.28.8
python-dotenv==1.0.0
requests==2.31.0
